Made a separate "ItemArray" classes out of a "ListBox.ItemArray" subclasses
